Move EndpointHostHelpers to the Aspire.Hosting.ApplicationModel namespace.
