+++ tests/Aspire.Hosting.Tests/ApplicationModel/EndpointHostHelpersTests.cs
@@ -1,9 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using Aspire.Hosting.Utils;
-
-namespace Aspire.Hosting.Tests.Utils;
+namespace Aspire.Hosting.Tests.ApplicationModel;
 
 public class EndpointHostHelpersTests
 {
